requirements.txt
Django==2.2.13
django-braces==1.14.0
django-crispy-forms==1.9.2
python-dateutil==2.8.1
ipinfo==3.0.0
